Fixed population report at long last
[skip actions]

+++ src/main/java/com/napier/sem/PopulationRecord.java
@@ -1,3 +1,3 @@
 package com.napier.sem;
 
-public record PopulationRecord(String name, int totalPopulation, int urbanPopulation, int ruralPopulation) {}
+public record PopulationRecord(String name, long totalPopulation, long urbanPopulation, long ruralPopulation) {}
